test(RoomescapeServiceTest): 서비스 분리
