+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -13,7 +13,6 @@ import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
 import roomescape.exception.exception.DataNotFoundException;
-import roomescape.exception.exception.DeletionNotAllowedException;
 import roomescape.exception.exception.DuplicateReservationException;
 import roomescape.exception.exception.PastReservationTimeException;
 
@@ -23,14 +22,18 @@ class ReservationServiceTest {
 
     @Autowired
     ReservationService service;
+    @Autowired
+    ReservationTimeService timeService;
+    @Autowired
+    ReservationThemeService themeService;
 
     @DisplayName("같은 날짜 및 시간 예약이 존재하면 예외를 던진다")
     @Test
     void addReservationWithDuplicatedReservation() {
         //given
         LocalDate date = LocalDate.now().plusDays(1);
-        
-        ReservationTimeResponse response = service.addReservationTime(
+
+        ReservationTimeResponse response = timeService.addReservationTime(
                 new ReservationTimeRequest(LocalTime.parse("10:10")));
 
         service.addReservation(new ReservationRequest("제프리", date, 1L, response.timeId()));
@@ -47,7 +50,7 @@ class ReservationServiceTest {
     @Test
     void addReservationBeforeCurrentDateTime() {
         // given
-        service.addReservationTime(new ReservationTimeRequest(LocalTime.parse("10:10")));
+        timeService.addReservationTime(new ReservationTimeRequest(LocalTime.parse("10:10")));
         LocalDate date = LocalDate.now().minusDays(1);
         ReservationRequest request = new ReservationRequest("호떡", date, 1L, 1L);
 
@@ -69,51 +72,4 @@ class ReservationServiceTest {
                 .hasMessage("[ERROR] 예약번호 999번에 해당하는 예약이 없습니다.");
     }
 
-    @DisplayName("존재하지 않는 예약시간을 삭제하려는 경우 예외를 던진다")
-    @Test
-    void removeReservationTime() {
-        //given
-        long notExistId = 999;
-
-        //when & then
-        assertThatThrownBy(() -> service.removeReservationTime(notExistId))
-                .isInstanceOf(DataNotFoundException.class)
-                .hasMessage("[ERROR] 예약 시간 999번에 해당하는 시간이 없습니다.");
-    }
-
-    @DisplayName("존재하지 않는 예약테마를 삭제하려는 경우 예외를 던진다")
-    @Test
-    void removeReservationTheme() {
-        //given
-        long notExistId = 999;
-
-        //when & then
-        assertThatThrownBy(() -> service.removeReservationTheme(notExistId))
-                .isInstanceOf(DataNotFoundException.class)
-                .hasMessage("[ERROR] 예약 테마 999번에 해당하는 테마가 없습니다.");
-    }
-
-    @DisplayName("예약시간과 연결된 예약이 존재하는 경우 예약시간을 삭제할 수 없다.")
-    @Test
-    void removeReservationTimeWithExistsReservation() {
-        // given
-        long existId = 1L;
-
-        // when & then
-        assertThatThrownBy(() -> service.removeReservationTime(existId))
-                .isInstanceOf(DeletionNotAllowedException.class)
-                .hasMessage("[ERROR] 예약이 연결된 시간은 삭제할 수 없습니다. 관련 예약을 먼저 삭제해주세요.");
-    }
-
-    @DisplayName("예약테마와 연결된 예약이 존재하는 경우 예약테마를 삭제할 수 없다.")
-    @Test
-    void removeReservationThemeWithExistsReservation() {
-        // given
-        long existId = 1L;
-
-        // when & then
-        assertThatThrownBy(() -> service.removeReservationTheme(existId))
-                .isInstanceOf(DeletionNotAllowedException.class)
-                .hasMessage("[ERROR] 예약이 연결된 테마는 삭제할 수 없습니다. 관련 예약을 먼저 삭제해주세요.");
-    }
 }
